requirements.txt
Pillow
beautifulsoup4